Backend/Data Proceesing and Search/requirements.txt
beautifulsoup4==4.11.1
blurhash==1.1.4
blurhash_python==1.1.3
firebase_admin==5.3.0
meilisearch==0.22.1
pandas==1.5.1
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
